gomod: update zoekt to fix List optimization
After investigating the performance impact of the List optimization, we
noticed we did not in practice use the performance improvement.
Additionally, in a large number of cases the more advanced
synchronization was leading to worse performance. So this update
includes two commits which fix the optimization and revert the
concurrency changes.

https://github.com/sourcegraph/zoekt/compare/931974dd953e...8cf8887a903a

- 7621acc2c5 shards: selectRepoSet supports queries which are not wrapped in and
- 8cf8887a90 Revert "shards: respect scheduler and use smarter synchronization for List"

Test Plan: tested in zoekt and this CI