gomod: update zoekt to include List performance improvement
NOTE: This is not a backport from sourcegraph's main branch, but rather
updating backports that have occurred on the zoekt sourcegraph-5.3
branch.

This includes an improvement performance improvement for List. This
change was generated with "./dev/zoekt/update sourcegraph-5.3". The
sourcegraph-5.3 branch tracks the version of zoekt used by the 5.3
release branch in sourcegraph.

https://github.com/sourcegraph/zoekt/compare/245e0cebf761...6364497e6146

- 8f6d308e8e shards: use selectRepoSet in List
- 2da0b9825a shards: respect scheduler and use smarter synchronization for List
- 31adf4c3db shards: selectRepoSet supports queries which are not wrapped in and
- 6364497e61 Revert "shards: respect scheduler and use smarter synchronization for List"

Test Plan: tested in zoekt repo and in this CI