embeddings: don't schedule jobs if disabled
We return early, either if embeddings are disabled or if there are no
repos matching any policy.

Before, we scheduled embeddings for all repos returned by
`ListMinimalRepos(ctx, database.ReposListOptions{IDs: nil})` in
`ScheduleRepositoriesForPolicy`. Reading the documentation
of `RepoListOptions.IDs`  "When zero-valued, this is omitted from the
predicate set", this might cause repo to get embedded even if embeddings
are disabled a no policies are defined.

Test plan:
CI