Revert "Remove repoRefresh endpoint (#57730)"
This reverts commit fd1363870613bd88453c38c27a16b5ba09c55d77.

A customer is making use of this, and until we actually work on the scheduler, it doesn't hurt having it, so we can make it work for them again for now and will find a better solution while we tackle the scheduler as a whole.

## Test plan

Manually hit the endpoint locally.