update tests which hardcore regex
Some tests hardcode the output of zoekt's query to string. The last
commit made use use the regex format from go1.21. This updates those
values to use them.

Additionally we switch a test to using autogold so that when we have
correctly workedaround this in the future updating will be easy.

Test Plan: CI