svelte/ci: Run `svelte-check` in CI
This commit updates the bazel build file to execute `svelte-check` in
CI.

To make this work correctly we need to run additional commands to
generate all the files that are needed by the code. Specifically that's
`svelte-kit sync` and the graphql code/type generation. The code was
refactored to make that work.

This commit also fixes issues detected by `svelte-check`.

IMPORTANT
Running `svelte-check` in CI differs in two ways from running it
locally:

1. Locally `svelte-check` will also check all local `@sourcegraph/*`,
   because they symlink the orginal source code.
   In CI/Bazel those dependencies are already built and therefore are
   not checked by Bazel. However, this is what we want because the
   dependencies do not always confirm to the rules set by
   `svelte-check`.
2. Source files that are not required for the production build have been
   excluded and are therefore not checked. That includes Storybook story
   files, playwright and vitest files and `src/testing/*`.