Add support for naming repo explicitly for Bitbucket Cloud
This aligns the configuration for Bitbucket Cloud with most other code hosts where we allow to name repos explicitly. This makes it easier to test the connection for a single repo first, instead of having to sync potentially thousands of repos right away.

## Test plan

Added a test using real-world data with VCR and verified manually on my local instance that I can successfully sync single repos.