Context: recognize more text extensions (#61573)
While digging into context quality, I noticed that for some types of question,
our approach of diversifying results across code and docs really helps.
However, we include a fairly narrow set of files as docs/ text files.

This PR extends the list of extensions we recognize. To come up with the new
list, I looked through the list of languages that GitHub linguist recognizes,
searched for type: prose, and included everything that was reasonably common
and unambiguous.