web: Switch to absolute blame recency calculation (#61661)
In a nutshell:

- We currently compute the recency color relative to either one year ago, three years ago, or the first commit of the repository.
- That means the colors have different meaning across different repositories.
- Fetching the date of the first commit can be expensive/slow.

This commit changes the method to calculate the color to always use the same time frame. The age of the commit is mapped to a number between [0, 1) and d3-scale-chromatic is used to
map that value to a color.
We can control the sensibility of the colors for newer/older commits by changing how we do the mapping. In this specific version I chose the midpoint to be ~ 6 months in the past.

I don't think there is "the right way" to perform the mapping but intuitively I would say that being able to better differentiate more recent commits is more useful than differentiating older commits.