gitserver: Don't reclone repos every 45 days
This PR simplifies the janitor so that it doesn't have to run clones out of band.

As a side-effect, this will prevent us to go over the configured concurrency for janitorial reasons. There are still a few other reasons but those will get eliminated later.

We used to reclone repos every 45 days, regardless of state. This is from the old dotcom, gitserver-is-only-a-cache days. In todays logic, this doesn't make a ton of sense since on most instances, all repos are expected to always be in a servable state.

Why don't we need this?

Many of the repos we keep don't actually need GC at all, so recloning is actually much more expensive for those. Especially for customers with many repos (and dotcom), it is expected that the majority of repos will get little to no changes.

In addition, we want to start generating additional git data structures like commitgraphs and file change bitmaps. Generating those is generally not free and they will be lost during a reclone.

So instead we now only try to detect when a repo is truly broken and will then delete it, so that the uncloned repos prioritizer will reclone them afterwards.

Test plan:

Adjusted tests for the janitor, and will monitor dotcom for a few days to see what impact this has.
If repos get too slow or the gc operations take too long, we can either revert or take additional measures.