gitserver: Remove RepoClone gRPC API
This was only used for a single purpose that can be achieved in a different way, but would provide a path around our scheduler which makes scheduling decisions harder to understand.
These clones would also not have stuck to the global concurrency limit.

Test plan:

Adjusted tests, tried a reclone locally and it worked.