gitserver: Simplify clone
After the PRs this is stacked on top of, the only option we passed to CloneRepo was Block: true.
This means that the in-memory queue was no longer used and could be removed.
This makes for a single code path to clone a repo, and a single caller on the gRPC side as well, so nice and simple.

This'll help make improvements to the scheduler simpler without having to think about 10 different call paths.

Test plan:

Just removed dead code, E2E tests still pass.