gitserver: Don't silently swallow errors from clone URL construction (#61587)
We currently just log warnings when failing to parse URLs and then return unparseable URLs which is set to explode later anyways. So this PR makes it more explicit and should make debugging this easier as well.

## Test plan

The test suite for clone URL still passes, integration and E2E tests verify that cloning still works as expected.