sg msp: add interactive version of 'sg msp init' and 'sg msp init-env' (#61697)
As titled - we can now prompt you through service setup. This is a small QOL improvement that makes providing the required parameters a bit easier, as the setup flags experience has been a point of feedback from several MSP adopters.

As a follow-up, we can probably remove the flag-based setup entirely, as it should generally be a human-operator-only setup. Then we can expand the setup to include e.g. resource setup (postgres), and maybe an initial generate step as well.

## Test plan

![image](https://github.com/sourcegraph/sourcegraph/assets/23356519/62810f24-bcce-4c8d-8bed-2f6c762ab45a)