search jobs: add suport for structural search
The adds support for strucural search queries to Search Jobs.

Limitations:

- No AND/OR queries, because structural search doesn't support `Basic` queries

This means

supported:
```
if ...
```

not supported
```
if ... OR switch ...
```

Notes:
- Structural search relies on `bufio.Scanner` to read stdout.
Potentially long lines might cause a search job to fail. In my local
tests it was very easy to hit that limit. We should make the buffer size
configurable (separate PR).

Test plan:
- Updated unit test