sgconf: allow TestParse{AndMerge} to run via go test
We don't have SG_FORCE_REPO_ROOT set, but we can just check the local
directory. We end up using the same logic I believe sg uses, but I don't
think this part of the assertations are the important part.

Alternatively we could just ignore this field in the cmp.Diff, but this
was the change of least resistance.

Test Plan: go test