Context: remove key words from stopwords (#61663)
While digging into why we perform poorly on "Ownership" and "Changelog"
searches, I noticed we were removing key terms in these queries. This PR adapts
the stopwords list to avoid removing them.

This doesn't immediately fix the problem, as we still don't match on filenames
(like CODEOWNERS or CHANGELOG.md). But it's a step towards the solution.