Context: match on filename (#61664)
Currently, Cody keyword context only matches on file contents. For many queries
it's critical to match the filename too. The reason we only matched contents is
purely historical -- I just chose the simplest implementation at first.

Examples of when this is important:
* "who owns third party licenses?" should match `third-party-licenses/CODEOWNERS`
* "grafana version deps.bzl" should clearly match `deps.bzl`

This PR extends the searches to match filename too. It doesn't immediately
improve results for most Ownership or Changelog queries, because Zoekt does not
count filename matches towards a result's score. I'll fix this in follow-up work.