monitoring: add golang monitoring for zoekt
Noticed this omission when I was wondering if we had goroutine leaks.
Our other services define this.

I added a simple way to indicate the container name in title since this
is the first service we added which needs this.

Test Plan: go test. Copy paste generated query into grafana explore on
dotcom.