svelte: Fix file view text selection issue
In the file view it looked like text selection doesn't work. However, I
did work but the decoration wasn't visible. That's because the text
selection decoration is rendered "underneath" the content, but the
content element has been given a background color, so the text selection
decoration wasn't visible.

The fix is to set the background color on the container that renders the
editor instead.