completions: recover from panics in writeEvent (#61754)
I timeboxed some time to work out why we call writeEvent once the
request is done. My main idea is around request cancellation and the
first event being an attribution error. However, I could not nail down
an exact cause. Instead this is a change to isolate panics to the
request it is happening in rather than taking down the whole process.

Test Plan: go test