Remove clone progress DB experiment (#61615)
This was an experiment from a while ago about storing the clone progress in the DB instead of making a request to gitserver to retrieve it. This has never been activated and is thus dead code, so removing it.
With the switch to a dbworker we get better logging for free anyways.

## Test plan

Removed dead code, tests should cover any regressions.