svelte: Refactor search cache URL parameter processing
While looking into
https://github.com/sourcegraph/sourcegraph/issues/61636 I thought we
might need to support more than two caching policies, which let me to
rewrite how handle this URL parameter.

It turned out that the root cause for the linked issue is something
different but I still think that the changed code introduces a nice
layer of abstraction.