gitserver: Remove double negation for ensure revision (#61676)
I _ALWAYS_ stumble over this double negation. "If not noensurerevision". huh?

Also, we should not encourage people to accidentally make the server do a resolve revision call which is generally costly, so the cheap option as the default seems more desirable.

For now, I haven't changed any of the values, but we should also do another review of the cases where we explicitly set it to true.

Test plan:

Just converted a bool, CI and E2E tests should catch any issues.