conf: Move most confvalidators into common package (#61810)
Previously, some of these validators would not fire correctly, because some packages aren't imported from cmd/frontend.
The package that contributes validators MUST be imported into frontend, which is not guaranteed. By colocating the GetWarnings func and the validators in one package, this is effectively enforced.

There are a few left that were a little harder to migrate, but this feels like a useful first step (and the end of my time window for this rn 😬)

Test plan:

Moved code around, my instance can still start fine and doesn't report new errors.