chore: Remove sentinel code (#61814)
This feature has not shipped, and is currently not prioritized. Git history is always preserved, but less code to actively maintain on `main` is always good, so I think this change is a win.

I am not the owner of this code, so I will defer to the graph team to decide if this should be merged or closed.

Test plan:

Code review to verify this change is ok, and CI to verify the code didn't break.