graphql: expose indexed search endpoints
This is part of the effort to enable admins to drain a Zoekt instance,
which will help to downscale Zoekt clusters without downtime.

The idea is that admins can query our GraphQL API to get a list of
available endpoints. In another PR, I will allow admins to list the
endpoints they want to drain.

Test plan:
manual testing