gitserver: Extract repository service in gRPC
This PR moves the Fetch and Delete operations onto a separate service.

With that, we also disallow consumers of the internal/gitserver.Client
interface to use these methods. Repo-updaters scheduler has authority
over this and it should not be interfered with. If a caller wants to make
a scheduling _suggestion_, it will be able to talk to repo-updater.

This prepares for a better scheduler that has full control of the operations
running.

Test plan:

E2E and integration tests will verify that repos are still cloned properly.