highlight: share log attributes with trace and handle canceled (#61880)
While investigating a large number of errors in highlighting I noticed
the vast majority where canceled. This commit adds special handling for
it and in the process does a few other things to the observability here:

- Log duration in these errors. This is useful signal to understand if a
  slow request is being canceled, or we have bad debouncing in a client.
- Re-use the tracers logger. This has many more fields and avoids
  duplication.
- Downgrade cancel error to warning.
- Include error problem in logs.
- Factor out the snippet log field creation.

Note I didn't include snippet in the cancel log since it's unlikely to
be related to the code being highlighted. Easy to add in if we find that
is incorrect.

Test Plan: CI will exercise this code path. Otherwise I will follow-up
on dotcom.

(cherry picked from commit b89b99c515087dc4273fbb911320934b39185908)