Svelte: implement basic git blame view (#61809)
This implements a basic blame view in the Svelte blob view. It is a simplified version of what we have in the React webapp, and notably does not include the popover. I think this is still usable as-is, and the PR was getting large, so I cut it off here, but I will update the issue with the followups.