gitserver: Migrate Diff to gRPC call
This PR migrates the Diff method to use a proper gRPC method. I decided to call it RawDiff as it streams back the raw diff format.

Later, we can decide to also add a structured endpoint, but for now we'll let the caller decide what to do with the diff string.
There's probably a use-case for both versions.

Closes https://github.com/sourcegraph/sourcegraph/issues/61868

Test plan:

Added tests across all the layers, verified locally that the diff view works.