gitserver: Completely remove Archive path validation
This didn't add a ton of benefits as the response will just be an empty archive and no error, but broke some callers of this, so we decided to remove it entirely.

This readds treeish validation, by using git cat-file -t -- <treeish> which works for refs, objects, trees, commits, ... which is probably slightly too permissive but should give good error codes for the common cases (and as far as I can tell no false negatives).

Test plan:

Existing tests pass, adjusted tests to no longer account for the not found case.