gitserver: Move subrepoperms checks back to client
While starting this migration, it felt nice to move sub repo permission checks to the server side, because that means we can make certain optimizations and none of the clients need to reimplement the SRP checker implementation.

However, this moves sub repo authz over the wire, but doesn't move repo perms generally here, so this was a weird split.

This also unnecessarily couples Sourcegraph and gitserver, so for now this is going back into the client layer, like all the auth checks.

Please make sure to pay good attention to the backend and client implementation and help me validate that those are indeed equal.

Test plan:

Added tests, E2E perforce subrepo perms tests are still passing.