appliance: reconciliation helpers (#61932)
* appliance: set pseudo-CRD name values

* appliance: receive and store requested version

Make the version part of the appliance-driven config, and store it on a
ConfigMap annotation. Service-specific upgrade logic then has access to
both the old and the new version, supporting complex multiversion
upgrades as needed.

The "Sourcegraph" config object is a kubebuilder-scaffolded custom type,
but we don't actually use CRDs. We drive its status field using this
ConfigMap annotation, so that lower-level code can behave as if we do
use CRDs. This is similar to what we do with the namespace field.

* appliance: add some standard labels to default deployment

Notably the version, which might be useful for debugging. Label keys
lifted from the helm chart.

* appliance: deploy blobstore deployment using new reconciliation logic