gitserver: More sophisticated argument building
This PR aims to improve our command building and validation. Previously, every internal command would need to be added to the command allow list.

This meant that whatever we allow internal actors to do, we would also need to allow callers of exec to do, which exposes some risk.

This PR aims to clean up with that by marking certain arguments that we hard-code as safe, and verifies the other ones for safety.

Test plan:

E2E and integration tests are still passing, requesting review from security.