gitserver: Simplify invocation of p4-fusion
This wrapper existed to track performance issues, but we have since added better logging options for that through storing the last output and the ability to log to a file.

This adds another layer of indirection that could cause trouble, so simplifying here and calling p4-fusion directly, like we do in the p4-fusion CI pipeline.

Test plan:

The p4 integration test still passes.