[Backport 5.3.9104] Context: avoid panic on stopwords query (#62026) (#62055)
* Context: return no results for stopwords query (#61848)

In #60106, we fixed a bug where the context search threw an error when the query
was composed entirely of stopwords. However, the queries were still returning
results (based on filename match). This means that for a stopwords-only query
like "what's going on?" we return a fairly random set of files, which confuses
the LLM.

Now we make sure to return an empty search, which returns no results.

* Context: avoid panic on stopwords query (#62026)

In #61848, we tried to fix an issue when the query contained only stopwords and
we interpreted this as "match all files". Unfortunately this fix introduced a panic by
returning a nil search job, which is not allowed by the our job framework.

Now, we return a `noopJob` which returns no results. This is the same approach
we use when an AND/ OR query has no operands.